Added Simple block Structure and Creative Tab
Implements a simple structure for adding non-tile blocks. TODO: Add model and texture system.

Also adds a creative tab for when we start adding items to the game. This should allow easier testing than using chat commands to spawn everything in.

Lastly, there is now a language file. Look at it for reference of how to name things in game.